catalog-backend: inline synthetic load test module
